pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orrbeam"
version = "0.1.0"
description = "Unified Sunshine/Moonlight mesh — bidirectional remote desktop nodes"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "zeroconf>=0.131",
    "cryptography>=42.0",
]

[project.scripts]
orrbeam = "orrbeam.cli:main"
orrbeamd = "orrbeam.daemon:main"

[tool.ruff]
target-version = "py310"
line-length = 120
